Large commit involving multiple edits and additions
Added two new M-files, added multiple citations to the Bib file, and made edits to Prospectus and Prospectus Presentation TeX files. Also updated the functionality of testFunction.m to no longer have dimension as an input. Changes will have to be made to some M-files to reflect this updated function.